LFS + FlatWorld fix
Former-commit-id: d39ebd4ff8b5ee00ec3be9cc384040519b927465 [formerly 02f68f59d446815c71fa00e77b69e6dff62cb428]
Former-commit-id: da7e03663170d9fa8cd186ac70d3b220adf8719a

+++ Plugins/Voxel/Source/Voxel/Private/VoxelWorldGenerators/PerlinNoiseWorldGenerator.cpp
@@ -57,11 +57,11 @@ FVoxelMaterial UPerlinNoiseWorldGenerator::GetDefaultMaterial(int X, int Y, int 
 	}
 	else if (Z < 0)
 	{
-		return FVoxelMaterial(0, 1, FMath::Clamp<uint8>((10 + Z) * 256.f / 10.f, 0, 255));
+		return FVoxelMaterial(0, 1, FMath::Clamp<int>((10 + Z) * 256.f / 10.f, 0, 255));
 	}
 	else if (Z < 10)
 	{
-		return FVoxelMaterial(2, 1, 255 - FMath::Clamp<uint8>(Z * 256.f / 10.f, 0, 255));
+		return FVoxelMaterial(2, 1, 255 - FMath::Clamp<int>(Z * 256.f / 10.f, 0, 255));
 	}
 	else if (Z < 100)
 	{
